+++ KursProjekt_2/camel.cpp
@@ -3,7 +3,7 @@
 #include "transport.h"
 #include < math.h >
 
-camel::camel(int a, double b, double c, double c1, double x)
+camel::camel(double a, double b, double c, double c1, double x)
 {
 	name = "Верблюд";
 	speed = a;
@@ -11,7 +11,7 @@ camel::camel(int a, double b, double c, double c1, double x)
 	rest[0] = c;
 	rest[1] = c1;
 	way = x;
-	time = way / speed;
-	if ((time / DrivingTime > 1 && time / DrivingTime < 2)) time = time + rest[0];
-	else if ((time / DrivingTime >= 2)) time = time + rest[0] + rest[1] * (ceil(time / DrivingTime) - 2);
+    time = way / speed;
+	if ((time / DrivingTime > 1 && time / DrivingTime < 2))  time = time + rest[0];
+	else if ((time / DrivingTime >= 2))  time = time + rest[0] + rest[1] * (ceil(time / DrivingTime) - 2);
 }
