+++ KursProjekt_2/camelFast.cpp
@@ -3,7 +3,7 @@
 #include "transport.h"
 #include < math.h >
 
-camelFast::camelFast(int a, double b, double c, double c1, double c2, double x) // Верблюд-быстроход
+camelFast::camelFast(double a, double b, double c, double c1, double c2, double x) // Верблюд-быстроход
 {
 	name = "Верблюд-быстроход";
 	speed = a;
@@ -12,8 +12,8 @@ camelFast::camelFast(int a, double b, double c, double c1, double c2, double x) 
 	rest[1] = c1;
 	rest[2] = c2;
 	way = x;
-	time = way / speed;
-	if ((time / DrivingTime > 1 && time / DrivingTime < 2)) time = time + rest[0];
-	else if ((time / DrivingTime > 2 && time / DrivingTime < 3)) time = time + rest[0] + rest[1];
-	else if ((time / DrivingTime >= 3)) time = time + rest[0] + rest[1] + rest[2] * (ceil(time / DrivingTime) - 3);
+    time = way / speed;
+	if ((time / DrivingTime > 1 && time / DrivingTime < 2))  time = time + rest[0];
+	else if ((time / DrivingTime > 2 && time / DrivingTime < 3))  time = time + rest[0] + rest[1];
+	else if ((time / DrivingTime >= 3))  time = time + rest[0] + rest[1] + rest[2] * (ceil(time / DrivingTime) - 3);	
 };
